am62px: SOC: Fix Misra C Rule 12.1 issue
MISRA.EXPR.PARENS

The precedence of operators within expressions should be made explicit.

Fixes: SITSW-7018

+++ source/drivers/soc/am62px/soc.c
@@ -452,7 +452,7 @@ void SOC_controlModuleLockMMR(uint32_t domainId, uint32_t partition)
     {
         /* in AM62Ax, main dowmin MMRs are left unlocked since when working with linux kernel, linux kernel assumes MMRs are unlocked */
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_CTRL_MMR0_CFG0_BASE);
-        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)(CSL_MAIN_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
+        kickAddr = (volatile uint32_t *) ((uint32_t)baseAddr + ((uint32_t)CSL_MAIN_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
         CSL_REG32_WR(kickAddr, KICK_LOCK_VAL);      /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK_LOCK_VAL);      /* KICK 1 */
@@ -461,7 +461,7 @@ void SOC_controlModuleLockMMR(uint32_t domainId, uint32_t partition)
     if((SOC_DOMAIN_ID_MCU == domainId) || (SOC_DOMAIN_ID_WKUP == domainId))
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_MCU_CTRL_MMR0_CFG0_BASE);
-        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)(CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
+        kickAddr = (volatile uint32_t *) ((uint32_t)baseAddr + ((uint32_t)CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
         CSL_REG32_WR(kickAddr, KICK_LOCK_VAL);      /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK_LOCK_VAL);      /* KICK 1 */
@@ -470,7 +470,7 @@ void SOC_controlModuleLockMMR(uint32_t domainId, uint32_t partition)
     if(SOC_DOMAIN_ID_WKUP == domainId)
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_WKUP_CTRL_MMR0_CFG0_BASE);
-        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)(CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
+        kickAddr = (volatile uint32_t *) ((uint32_t)baseAddr + ((uint32_t)CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
         CSL_REG32_WR(kickAddr, KICK_LOCK_VAL);      /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK_LOCK_VAL);      /* KICK 1 */
@@ -488,7 +488,7 @@ void SOC_controlModuleUnlockMMR(uint32_t domainId, uint32_t partition)
     if(SOC_DOMAIN_ID_MAIN == domainId)
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_CTRL_MMR0_CFG0_BASE);
-        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)(CSL_MAIN_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
+        kickAddr = (volatile uint32_t *) ((uint32_t)baseAddr + ((uint32_t)CSL_MAIN_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
@@ -497,7 +497,7 @@ void SOC_controlModuleUnlockMMR(uint32_t domainId, uint32_t partition)
     if(SOC_DOMAIN_ID_MCU == domainId)
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_MCU_CTRL_MMR0_CFG0_BASE);
-        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)(CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
+        kickAddr = (volatile uint32_t *) ((uint32_t)baseAddr + ((uint32_t)CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
@@ -506,7 +506,7 @@ void SOC_controlModuleUnlockMMR(uint32_t domainId, uint32_t partition)
     if(SOC_DOMAIN_ID_WKUP == domainId)
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_WKUP_CTRL_MMR0_CFG0_BASE);
-        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)(CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
+        kickAddr = (volatile uint32_t *) ((uint32_t)baseAddr + ((uint32_t)CSL_MCU_CTRL_MMR_LOCKn_KICK0_OFFSET(partition)));
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
